Network.Linklater.Types

Documentation

type ChannelID = Text

The unique 'C<number>' Slack assigns to each channel. Used to say things.

data Channel

Where slash commands come from and where Messages go.

Constructors

Channel !ChannelID !Text

newtype User

A username: no at-signs, just text!

Constructors

User Text

data Command

Incoming HTTP requests to the slash function get parsed into one of these babies.

Constructors

Command

Fields

_commandName :: !Text

The command name.

_commandUser :: !User

Who ran your slash command.

_commandChannel :: !Channel

Where the person ran your slash command.

_commandText :: !(Maybe Text)

Text for the slash command, if any.

newtype Icon

The icon next to the messages you say. (Images unsupported right now, sorry.)

Constructors

EmojiIcon Text
For example, ":stars2:".

data Format

A little DSL for Slack formatting.

Constructors

FormatAt !User
"<@user|user>"

FormatUser !User !Text
"<@user|user did this and that>"

FormatLink !Text !Text
"<http://example.com|user did this and that>"

FormatString !Text
"user did this &amp; that"

data Message

Here's how you talk: you make one of these and pass it to say. Before the day is done, Linklater will convert this to a JSON blob using Aeson.

Simple messages are parsed by Slack with parse=full (i.e. as if you had typed it into the input box).

Complex messages are parsed according to Slack formatting. See Format.

Constructors

SimpleMessage !Icon !Text !Channel !Text

FormattedMessage !Icon !Text !Channel ![Format]

data Config

Like a curiosity about the world, you'll need one of these to say something.

Constructors

Config

Fields

_configHookURL :: !Text

This is the incoming web hook URL that Slack gave you. It's usually https://hooks.slack.com/services/....

newtype APIToken

An API token, either a tester token or one obtained through OAuth.

See:

https://api.slack.com/docs/oauth-test-tokens

https://api.slack.com/docs/oauth

Constructors

APIToken Text

type Response = Response ByteString

data RequestError

Rather than throwing unchecked exceptions from our HTTP requests, we catch all unchecked exceptions and convert them to checked RequestError values.

Constructors

RequestError

Fields

_requestErrorException :: !(Maybe SomeException)

An unchecked exception, typically IOException, occurred

_requestErrorResponse :: !(Maybe Response)

Something awry with the response, typically a non-2xx response

_requestErrorDecoding :: !(Maybe String)

Something awry with the JSON decoding

unformat :: Format -> Text

commandOfParams :: Map Text Text -> Either String Command

tryRequest :: (MonadIO m, MonadError RequestError m) => IO Response -> m Response

promoteEither :: MonadError RequestError m => Response -> (l -> String) -> Either l r -> m r

promoteMaybe :: MonadError RequestError m => Response -> String -> Maybe r -> m r

asList :: [a] -> [a]